add styling to newchat page and add a page that confirms a new chat request
