+++ src/App.css
@@ -10,42 +10,6 @@
 	font-family: "Fira sans", sans-serif;
 }
 
-/* box that contains topics on the information page */
-.infoBox {
-	background-color: var(--light);
-	color: var(--dark);
-	border: 2px solid var(--dark);
-	margin: 10px;
-	padding: 5px;
-	border-radius: 5px;
-}
-
-/* box that contains the log-out functionality */
-.logoutBox {
-    /*make this box less wide and centralise*/
-    background-color: var(--light);
-	color: var(--dark);
-	border: 2px solid var(--dark);
-	margin: 10px;
-	padding: 5px;
-	border-radius: 5px;
-}
-
-
-
-/* section that holds topic tiles */
-.topics {
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: row;
-    justify-content: space-between;
-    width: calc(100% - 40px);
-    padding: 10px;
-    margin: 10px;
-    padding: 5px;
-    overflow-y: auto; 
-}
-
 /* general button */
 .button {
     background-color: #ffffff;
@@ -59,26 +23,4 @@
     margin: 4px 4px;
     cursor: pointer;
 
-}
-
-/* button that contains topics on New-Chat Page */
-.topicButton {
-    border-radius: 40px;
-}
-
-.topicButtonclicked {
-    border-radius: 40px;
-    background-color: #000000;
-    color: rgb(255, 255, 255);
-}
-
-/* button that requests new chats */
-.requestbutton {
-    background-color: #000000;
-    color: rgb(255, 255, 255);
-}
-
-/* button used to log out */
-.logoutbutton {
-    font-size: 10px;
 }
