+++ src/components/pages/index.jsx
@@ -1,3 +1,4 @@
 export * from "./Chats";
 export * from "./Information";
 export * from "./NewChat";
+export * from "./NewChatRequested";
